feat(ui): add version number to header
