+++ src/routes/Footer.svelte
@@ -1,0 +1,13 @@
+<script lang="ts">
+	import { AppBar } from '@skeletonlabs/skeleton';
+</script>
+
+<AppBar gridColumns="grid-cols-1" slotDefault="place-self-center">
+	<p class="text-center text-sm">
+		Any issues or suggestions? Please get in touch!
+		<br />
+		<a href="mailto:smesnage@sheffield.ac.uk">
+			<strong>smesnage@sheffield.ac.uk</strong>
+		</a>
+	</p>
+</AppBar>
